requirements.txt
# Core dependencies
requests>=2.31.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
reportlab>=4.0.7

# Google Cloud dependencies
google-cloud-aiplatform>=1.38.1
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.13.0
google-cloud-secret-manager>=2.16.4
google-cloud-logging>=3.8.0
google-cloud-monitoring>=2.16.0

# Additional dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3

# AI and ML
vertexai==0.0.1
google-generativeai==0.3.2

# Utilities
typing-extensions==4.8.0 